Move graph collectors to graph walkers module
